python-backend/requirements.txt
fastapi[all]==0.112.1
uvicorn==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
bcrypt==3.2.0
passlib==1.7.4
python-jose==3.3.0
pydantic==2.8.2